Improve styling of projects links
To make the page more presentable, the corners of the container were rounded, padding was added and all the element in the container were centred. The images were given a border and their size limits were changed, so that the images would no longer have the width as its container.